Убрал функции, работа которых связана с форматированием данных в отдельный файл. Исправил название хука, который забирает данные для карточек.
